refactor(jmap): split mail/jmap.rs by domain (#166)
Pure move into mail/jmap/{mod,mail,mailboxes,identities,
calendar,contacts}.rs. Domain methods become per-file impl blocks
sharing the private api_request core; find_mailbox_by_role and
find_identity_id become pub(super) for cross-domain callers in
mail.rs. External crate::mail::jmap::* paths unchanged via
re-exports. update_calendar_event moved as-is; it has no callers.
